change homepage to guide homepage
