+++ src/index.tsx
@@ -7,10 +7,10 @@ import allReducers from "./reducers";
 import { Provider } from "react-redux";
 
 import App from "./App";
-import Homepage from "./pages/Homepage/Homepage";
+import Homepage from "./pages/AllPortfolio/AllPortfolio";
+import AllPortfolio from "./pages/AllPortfolio/AllPortfolio";
 import AllResumes from "./pages/AllResumes/AllResumes";
 import Profile from "./pages/Profile/Profile";
-import FollowingArea from "./pages/Following/FollowingArea";
 import Resume from "./pages/Resume/Resume";
 import Website from "./pages/Website/Website";
 import Portfolio from "./pages/Portfolio/Portfolio";
@@ -28,6 +28,7 @@ root.render(
       <Routes>
         <Route path="/" element={<App />}>
           <Route index element={<Homepage />} />
+          <Route path="allresumes" element={<AllPortfolio />} />
           <Route path="allresumes" element={<AllResumes />} />
           <Route path="login" element={<Login />} />
           <Route path="profile/:id" element={<Profile />} />
